+++ moin2gitwiki/moin2markdown.py
@@ -1,3 +1,4 @@
+import re
 from pathlib import Path
 from typing import Optional
 
@@ -8,6 +9,14 @@ from sh import pandoc
 
 from .fetch_cache import FetchCache
 from .wikiindex import MoinEditEntry
+
+
+def is_a_linemark_para(tag):
+    return (
+        tag.name == "p"
+        and tag.has_attr("class")
+        and re.match(r"line\\d+", tag["class"][0])
+    )
 
 
 @attr.s(kw_only=True, frozen=True, slots=True)
@@ -21,6 +30,41 @@ class Moin2Markdown:
     #
     # prebuilt sh setup
     pandoc = pandoc.bake("-f", "html", "-t", "gfm")
+    #
+    # smiley mapping
+    smiley_map = {
+        "X-(": ":rage:",
+        ":(": ":confused:",
+        ";)": ":wink:",
+        ":-?": ":stuck_out_tongue:",
+        ":-(": ":frowning_face:",
+        ";-)": ":wink:",
+        "{X}": ":x:",
+        "{3}": ":three:",
+        ":D": ":grin:",
+        ":)": ":slightly_smiling_face:",
+        "/!\\": ":warning:",
+        ":\\": ":confounded:",
+        ":-)": ":smiley:",
+        "|-)": ":pensive:",
+        "{i}": ":information_source:",
+        "{*}": ":star:",
+        "<:(": ":mask:",
+        "B)": ":sunglasses:",
+        "<!>": ":warning:",
+        ">:>": ":imp:",
+        "B-)": ":nerd_face:",
+        "(./)": ":white_check_mark:",
+        "{1}": ":one:",
+        "{o}": ":star:",
+        ":o": ":anguished:",
+        ":))": ":rofl:",
+        "(!)": ":bulb:",
+        "|)": ":monocle_face:",
+        ":-))": ":rofl:",
+        "{OK}": ":ok:",
+        "{2}": ":two:",
+    }
 
     @classmethod
     def create_translator(
@@ -65,6 +109,31 @@ class Moin2Markdown:
         # now strip out excess rubbish - anchor spans
         for tag in content.find_all(class_="anchor"):
             tag.decompose()
+        #
+        # Remove dead <p class="line???"> with no closer
+        for tag in content.find_all(is_a_linemark_para):
+            tag.unwrap()
+        #
+        # now find all the links, and if within the wiki, rewrite
+        for tag in content.find_all("a"):
+            target = tag["href"]
+            if target:
+                url = self.url_prefix.copy().join(target)
+                if url.url.startswith(self.url_prefix.url):
+                    new_url = url.url[len(self.url_prefix.url) :]
+                    if new_url in self.link_table:
+                        tag["href"] = self.link_table[new_url]
+            #
+            # strip any class attributes on links - tend to upset the translator
+            if tag.has_attr("class"):
+                del tag["class"]
+        #
+        # now find all the images and see if they map to emojis
+        # MoinMoin puts the emoji code in the title, so will purely match on that
+        for tag in content.find_all("img"):
+            if tag.has_attr("title") and tag["title"] in self.smiley_map:
+                tag.replace_with(" " + self.smiley_map[tag["title"]] + " ")
+
         return content.contents
 
     def translate(self, input: str) -> str: